feat: :construction: Separate `check_pending_forest_root_writes` into MSP and BSP specific functions
